+++ R/rmonad.R
@@ -28,7 +28,8 @@ NULL
 #'
 #' @section Advanced operators:
 #'
-#'  \%*>\%    bind lhs list as arguments to right
+#'  \%*>\%    bind lhs list as arguments to right. The lhs may be a literal
+#'            list or a monad bound list.
 #'
 #'  \%^>\%    monadic bind and record input in monad
 #'
@@ -42,7 +43,7 @@ NULL
 #'
 #' as_monad - evaluate an expression into a monad (capturing error)
 #'
-#' lsmeval - evaluate expressions into a list inside a monad
+#' funnel - evaluate expressions into a list inside a monad
 #'
 #' @section monad to monad functions:
 #'
@@ -89,6 +90,6 @@ NULL
 #' as_monad(stop("instant death"))
 #'
 #' # convert multiple expressions into a list inside a monad
-#' lsmeval(stop("oh no"), runif(5), sqrt(-1))
+#' funnel(stop("oh no"), runif(5), sqrt(-1))
 
 NULL