+++ tests/testthat/test-binops.R
@@ -103,6 +103,12 @@ test_that('%*>% safely evaluates failing lists', {
   # but knows the input is failing
   expect_false( list(stop(1),5,2) %*>% max %>% m_OK )
 })
+test_that('%*>% works the same of monad bound lists', {
+  expect_equal(  funnel(3,5,2)       %*>% max %>% m_value, 5                )
+  expect_silent( funnel(stop(1),5,2) %*>% max                               )
+  expect_equal(  funnel(stop(1),5,2) %*>% max %>% m_value, list(NULL, 5, 2) )
+  expect_false(  funnel(stop(1),5,2) %*>% max %>% m_OK                      )
+})
 
 test_that('anonymous expressions can be run', {
   expect_equal( 1:10 %>>% { 4 } %>% esc  , 4 )